feat: added sidebar navigation component, used it on sidebar
